feat: employee info card add styles: renaming
